crowd_monitoring_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.24.3
selenium==4.15.2
python-multipart==0.0.6
